+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Date: 2025-06-XX
 
 ### Changed
 * Particle: The rapidity is now calculated with an `arcsinh` function to avoid numerical issues. If this is not possible, it falls back to the old method using the `log` function.
+* Histogram: The class does not throw a warning anymore if a value added to a histogram is outside the bin range. Instead, it simply ignores the value.
 * Tests: Test BulkObservables rapidity distribution with massive particles in a more realistic way.
 
 [Link to diff from previous version](https://github.com/smash-transport/sparkx/compare/v2.0.2...v2.1.0)